Add verification logic to success webhook
