+++ .env.example
@@ -14,6 +14,7 @@ DB_PORT_TEST=5433
 VIVA_WALLET_SOURCE_CODE="6343"
 VIVA_WALLET_CLIENT_CREDENTIALS="ZTc2cnBldnR1cmZma3RuZTduMTh2MG94eWozbTZzNTMycjFxNHk0azR4eDEzLmFwcHMudml2YXBheW1lbnRzLmNvbTpxaDA4RmtVMGRGOHZNd0g3NmpHQXVCbVdpYjlXc1A="
 VIVA_WALLET_VERIFICATION_KEY="94FA5D3BA6DBC79DC56E6BC7E2F8A3F25A566EAE"
+VIVA_WALLET_URL="https://demo-api.vivapayments.com"
 
 # Keycloak
 KEYCLOAK_CLIENT_ID=GoClient
