+++ app/config/config.go
@@ -10,6 +10,7 @@ type config struct {
 	VivaWalletSourceCode        string
 	VivaWalletClientCredentials string
 	VivaWalletVerificationKey   string
+	VivaWalletURL               string
 }
 
 // Config is the global configuration variable
@@ -21,6 +22,7 @@ var Config = config{
 	VivaWalletSourceCode:        getEnv("VIVA_WALLET_SOURCE_CODE", ""),
 	VivaWalletClientCredentials: getEnv("VIVA_WALLET_CLIENT_CREDENTIALS", ""),
 	VivaWalletVerificationKey:   getEnv("VIVA_WALLET_VERIFICATION_KEY", ""),
+	VivaWalletURL:               getEnv("VIVA_WALLET_URL", ""),
 }
 
 // Local copy of utils.GetEnv to avoid circular dependency
