--- app/paymentprovider/vivawallet_order.go
@@ -1,266 +1,0 @@
-package paymentprovider
-
-import (
-	"augustin/config"
-	"augustin/utils"
-	"bytes"
-	"encoding/json"
-	"errors"
-	"io"
-	"strconv"
-
-	"net/http"
-	"net/url"
-	"time"
-)
-
-var log = utils.GetLogger()
-
-type PaymentOrderRequest struct {
-	Amount              int      `json:"amount"`
-	CustomerTrns        string   `json:"customerTrns"`
-	Customer            Customer `json:"customer"`
-	PaymentTimeout      int      `json:"paymentTimeout"`
-	Preauth             bool     `json:"preauth"`
-	AllowRecurring      bool     `json:"allowRecurring"`
-	MaxInstallments     int      `json:"maxInstallments"`
-	PaymentNotification bool     `json:"paymentNotification"`
-	TipAmount           int      `json:"tipAmount"`
-	DisableExactAmount  bool     `json:"disableExactAmount"`
-	DisableCash         bool     `json:"disableCash"`
-	DisableWallet       bool     `json:"disableWallet"`
-	SourceCode          string   `json:"sourceCode"`
-	MerchantTrns        string   `json:"merchantTrns"`
-	Tags                []string `json:"tags"`
-	// TODO: Additional fields that Aaron found in the API docs
-	// PaymentMethodFees   []struct {
-	// 	PaymentMethodId int `json:"paymentMethodId"`
-	// 	Fee             int `json:"fee"`
-	// } `json:"paymentMethodFees"`
-	// CardTokens []string `json:"cardTokens"`
-}
-
-type Customer struct {
-	Email       string `json:"email"`
-	Fullname    string `json:"fullName"`
-	Phone       string `json:"phone"`
-	CountryCode string `json:"countryCode"`
-	RequestLang string `json:"requestLang"`
-}
-
-type AuthenticationResponse struct {
-	AccessToken string `json:"access_token"`
-	ExpiresIn   int    `json:"expires_in"`
-	TokenType   string `json:"token_type"`
-	Scope       string `json:"scope"`
-}
-
-type PaymentOrderResponse struct {
-	OrderCode int `json:"orderCode"`
-}
-
-type TransactionVerificationResponse struct {
-	Email               string  `json:"email"`
-	Amount              float64 `json:"amount"`
-	OrderCode           int     `json:"orderCode"`
-	StatusId            string  `json:"statusId"`
-	FullName            string  `json:"fullName"`
-	InsDate             string  `json:"insDate"`
-	CardNumber          string  `json:"cardNumber"`
-	CurrencyCode        string  `json:"currencyCode"`
-	CustomerTrns        string  `json:"customerTrns"`
-	MerchantTrns        string  `json:"merchantTrns"`
-	TransactionTypeId   int     `json:"transactionTypeId"`
-	RecurringSupport    bool    `json:"recurringSupport"`
-	TotalInstallments   int     `json:"totalInstallments"`
-	CardCountryCode     string  `json:"cardCountryCode"`
-	CardIssuingBank     string  `json:"cardIssuingBank"`
-	CurrentInstallment  int     `json:"currentInstallment"`
-	CardUniqueReference string  `json:"cardUniqueReference"`
-	CardTypeId          int     `json:"cardTypeId"`
-}
-
-func AuthenticateToVivaWallet() (string, error) {
-	// Create a new request URL using http
-	apiURL := "https://demo-accounts.vivapayments.com"
-	resource := "/connect/token"
-	jsonPost := []byte(`grant_type=client_credentials`)
-	u, _ := url.ParseRequestURI(apiURL)
-	u.Path = resource
-	urlStr := u.String() // "https://demo-accounts.vivapayments.com/connect/token"
-
-	req, err := http.NewRequest("POST", urlStr, bytes.NewReader(jsonPost))
-	if err != nil {
-		log.Error("Building request failed: ", err)
-	}
-
-	// Create Header
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Set("Authorization", "Basic "+config.Config.VivaWalletClientCredentials)
-	if config.Config.VivaWalletClientCredentials == "" {
-		err := errors.New("VivaWalletClientCredentials not in .env or empty")
-		log.Error(err)
-		return "", err
-	}
-
-	// Create a new client with a 10 second timeout
-	client := http.Client{Timeout: 10 * time.Second}
-	// Send the request
-
-	res, err := client.Do(req)
-	if err != nil {
-		log.Error("impossible to send request: ", err)
-	}
-
-	// Close the body after the function returns
-	defer res.Body.Close()
-	// Log the request body
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		log.Error("Reading body failed: ", err)
-		return "", err
-	}
-
-	// Unmarshal response body to struct
-	var authResponse AuthenticationResponse
-	err = json.Unmarshal(body, &authResponse)
-	if err != nil {
-		log.Error("Unmarshalling body failed: ", err)
-		return "", err
-	}
-
-	return authResponse.AccessToken, nil
-}
-
-func CreatePaymentOrder(accessToken string, amount int) (int, error) {
-	// Create a new request URL using http
-	apiURL := "https://demo-api.vivapayments.com"
-	resource := "/checkout/v2/orders"
-	u, _ := url.ParseRequestURI(apiURL)
-	u.Path = resource
-	urlStr := u.String() // "https://demo-accounts.vivapayments.com/connect/token"
-
-	// Create a new sample customer
-	// TODO: Change this to a real customer
-	customer := Customer{
-		Email:       "test@example.com",
-		Fullname:    "Test Customer",
-		Phone:       "1234567890",
-		CountryCode: "GR",
-		RequestLang: "en-GB",
-	}
-
-	// Create a new sample payment order
-	// TODO: Change this to a real payment order
-	paymentOrderRequest := PaymentOrderRequest{
-		Amount:              amount,
-		CustomerTrns:        "Augustin Straßenzeitung + Trinkgeld + Spende + Abo + Versand + 1x Ausgabe + probleme wenn der Text zu lang wird",
-		Customer:            customer,
-		PaymentTimeout:      300,
-		Preauth:             false,
-		AllowRecurring:      false,
-		MaxInstallments:     0,
-		PaymentNotification: true,
-		TipAmount:           0,
-		DisableExactAmount:  false,
-		DisableCash:         true,
-		DisableWallet:       true,
-		SourceCode:          utils.GetEnv("VIVA_WALLET_SOURCE_CODE", ""),
-		MerchantTrns:        "testMerchantTrns",
-		Tags:                []string{"testTag1", "testTag2"},
-	}
-
-	// Create a new post request
-	jsonPost, err := json.Marshal(paymentOrderRequest)
-	if err != nil {
-		log.Error("Marshalling payment order failed: ", err)
-	}
-
-	req, err := http.NewRequest("POST", urlStr, bytes.NewReader(jsonPost))
-	if err != nil {
-		log.Error("Building request failed: ", err)
-	}
-	// Create Header
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+accessToken)
-
-	// Create a new client with a 10 second timeout
-	client := http.Client{Timeout: 10 * time.Second}
-	// Send the request
-	res, err := client.Do(req)
-	if err != nil {
-		log.Error("impossible to send request: ", err)
-	}
-
-	if res.StatusCode != 200 {
-		return 0, errors.New("Request failed instead received this response status code: " + strconv.Itoa(res.StatusCode))
-	}
-
-	// Close the body after the function returns
-	defer res.Body.Close()
-	// Log the request body
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		log.Error("Reading body failed: ", err)
-		return 0, err
-	}
-
-	// Unmarshal response body to struct
-	var orderCode PaymentOrderResponse
-	err = json.Unmarshal(body, &orderCode)
-	if err != nil {
-		log.Error("Unmarshalling body failed: ", err)
-		return 0, err
-	}
-
-	return int(orderCode.OrderCode), nil
-
-}
-
-func VerifyTransactionID(accessToken string, TransactionID string) (success bool, err error) {
-	// Create a new request URL using http
-	apiURL := "https://demo-api.vivapayments.com"
-	resource := "/checkout/v2/transactions/" + TransactionID
-	u, _ := url.ParseRequestURI(apiURL)
-	u.Path = resource
-	urlStr := u.String() // "https://demo-api.vivapayments.com/checkout/v2/transactions/{TransactionID}"
-
-	// Create a new get request
-	req, err := http.NewRequest("GET", urlStr, nil)
-	if err != nil {
-		log.Error("Building request failed: ", err)
-	}
-	// Create Header
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+accessToken)
-
-	// Create a new client with a 10 second timeout
-	client := http.Client{Timeout: 10 * time.Second}
-	// Send the request
-	res, err := client.Do(req)
-	if err != nil {
-		log.Error("Sending request failed: ", err)
-	}
-
-	if res.StatusCode != 200 {
-		return false, errors.New("Request failed instead received this response status code: " + strconv.Itoa(res.StatusCode))
-	}
-
-	// Close the body after the function returns
-	defer res.Body.Close()
-	// Log the request body
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		log.Error("Reading body failed: ", err)
-		return false, err
-	}
-
-	// Unmarshal response body to struct
-	var transactionVerificationResponse TransactionVerificationResponse
-	err = json.Unmarshal(body, &transactionVerificationResponse)
-	if err != nil {
-		log.Error("Unmarshalling body failed: ", err)
-		return false, err
-	}
-	return true, nil
-}
